requirements.txt
streamlit>=1.18
yfinance
pandas
numpy
scikit-learn
ta
requests
python-dotenv
vaderSentiment
joblib
plotly
